fix a type mismatch and add finishing touches

+++ src/main/java/com/ps/FutureValue.java
@@ -7,7 +7,7 @@ public class FutureValue {
     float deposit;
     float interestPercent;
     float interestDecimal;
-    float numYears;
+    int numYears;
     final double daysInYear = 365;
     FutureValue() {
         // Receive user input data
